requirements.txt
streamlit>=1.32.0
python-docx>=0.8.11
google-generativeai>=0.3.0
